flask/requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
uuid==1.30
Werkzeug==3.0.2
